backend/requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
django_filter
drf-pyotp
qrcode==7.4.2
django-cors-headers==4.2.0
python-decouple==3.8
django-extensions
python-dotenv==1.0.0